[DataTypes] Make floating-point types explicit
- Fix bug in `save_pc_dat` / `load_pc_dat` to treat data as f32
- Make sure to use f32 everywhere except for poses, which are f64
- Anotate nv-maglev conversion with explicit types where necessary

+++ src/py/common/transformations.py
@@ -91,7 +91,7 @@ def euler_2_so3(euler_angles, degrees=True, seq='xyz'):
         (np array): rotations given so3 matrix representation [n,3,3]
     '''
 
-    return R.from_euler(seq=seq, angles=euler_angles, degrees=degrees).as_matrix()
+    return R.from_euler(seq=seq, angles=euler_angles, degrees=degrees).as_matrix().astype(np.float32)
 
 
 def axis_angle_2_quaternion(axis, angle, degrees=True):
